RecyclerView Funcionando!

+++ app/src/main/java/br/com/gabrieldargas/personalgym/models/Exercicio.kt
@@ -1,9 +1,11 @@
 package br.com.gabrieldargas.personalgym.models
 
+import com.google.firebase.firestore.Exclude
+
 
 data class Exercicio (
     val nome: String = "",
-    val series: Int = 0,
-    val repeticoes: Int = 0,
+    val series: Long = 0,
+    val repeticoes: Long = 0,
     var userId: String = ""
 )
